refactor(arcan): promote estimate_cost from shell.rs to shared arcan::cost module
The arcan agent test --live path (BRO-1008) needs the same USD cost
estimation the shell REPL uses for its summary line. Move the private
shell.rs helper (BRO-364) plus its unit tests into the library half of
the crate (arcan::cost) so both surfaces share one pricing table
instead of duplicating it.
